+++ pkgs/jnigen/lib/src/bindings/linker.dart
@@ -27,9 +27,8 @@ class Linker extends Visitor<Classes, Future<void>> with TopLevelVisitor {
   @override
   Future<void> visit(Classes node) async {
     void assignPath(ClassDecl decl) {
-      final root = config.outputConfig.dartConfig.path;
-      if (config.outputConfig.dartConfig.structure ==
-          OutputStructure.singleFile) {
+      final root = config.output.dart.path;
+      if (config.output.dart.structure == OutputStructure.singleFile) {
         decl.path = root.toFilePath();
       } else {
         final dollarSign = decl.binaryName.indexOf('\$');
@@ -68,7 +67,7 @@ class Linker extends Visitor<Classes, Future<void>> with TopLevelVisitor {
             config.importedClasses[binaryName] ?? node.decls[binaryName];
         if (decl != null) return decl;
 
-        if (config.generateStubs) {
+        if (config.output.generateStubs) {
           log.fine('Class $binaryName not found. Creating a stub.');
           // Create a synthetic stub. Mark it excluded for now. StubCollector
           // will mark it as a stub later.