+++ pkgs/jnigen/lib/src/elements/elements.dart
@@ -102,7 +102,7 @@ class ClassDecl with ClassMember, Annotated implements Element<ClassDecl> {
   @JsonKey(includeFromJson: false)
   BindingMode bindingMode;
 
-  bool get isExcluded => bindingMode == BindingMode.excluded;
+  bool get isIncluded => bindingMode != BindingMode.excluded;
   bool get isStub => bindingMode == BindingMode.stub;
 
   @JsonKey(includeFromJson: false)
@@ -608,7 +608,7 @@ class ArrayType extends ReferredType {
 mixin Annotated {
   abstract List<Annotation>? annotations;
 
-  static final nullableAnnotations = [
+  static const List<String> defaultNullableAnnotations = [
     // Taken from https://kotlinlang.org/docs/java-interop.html#nullability-annotations
     'org.jetbrains.annotations.Nullable',
     'org.jspecify.nullness.Nullable',
@@ -620,6 +620,10 @@ mixin Annotated {
     'lombok.Nullable',
     'io.reactivex.rxjava3.annotations.Nullable',
   ];
+  static final List<String> nullableAnnotations = [
+    ...defaultNullableAnnotations
+  ];
+
   bool get hasNullable {
     return annotations?.any(
           (annotation) =>
@@ -630,7 +634,7 @@ mixin Annotated {
         false;
   }
 
-  static final nonNullAnnotations = [
+  static const List<String> defaultNonNullAnnotations = [
     // Taken from https://kotlinlang.org/docs/java-interop.html#nullability-annotations
     'org.jetbrains.annotations.NotNull',
     'org.jspecify.nullness.NonNull',
@@ -642,6 +646,7 @@ mixin Annotated {
     'lombok.NonNull',
     'io.reactivex.rxjava3.annotations.NonNull',
   ];
+  static final List<String> nonNullAnnotations = [...defaultNonNullAnnotations];
   bool get hasNonNull {
     return annotations?.any(
           (annotation) =>
@@ -686,7 +691,7 @@ enum MethodKind {
 @JsonSerializable(createToJson: false)
 class Method with ClassMember, Annotated implements Element<Method> {
   Method({
-    this.userDefinedIsExcluded = false,
+    this.userDefinedIsIncluded = true,
     this.annotations,
     this.javadoc,
     this.modifiers = const {},
@@ -717,7 +722,7 @@ class Method with ClassMember, Annotated implements Element<Method> {
 
   /// Populated by user-defined visitors.
   @JsonKey(includeFromJson: false)
-  bool userDefinedIsExcluded;
+  bool userDefinedIsIncluded;
 
   /// Populated by user-defined visitors.
   @JsonKey(includeFromJson: false)
@@ -767,7 +772,7 @@ class Method with ClassMember, Annotated implements Element<Method> {
       returnType: returnType.clone(until: until),
       annotations: [...?annotations],
       descriptor: descriptor,
-      userDefinedIsExcluded: userDefinedIsExcluded,
+      userDefinedIsIncluded: userDefinedIsIncluded,
       javadoc: javadoc,
       modifiers: {...modifiers},
       params: params.map((param) => param.clone(until: until)).toList(),
@@ -805,7 +810,7 @@ class Method with ClassMember, Annotated implements Element<Method> {
       excluder:
       case GenerationStage.excluder:
       case GenerationStage.userVisitors:
-        cloned.userDefinedIsExcluded = userDefinedIsExcluded;
+        cloned.userDefinedIsIncluded = userDefinedIsIncluded;
         cloned.userDefinedName = userDefinedName;
       case GenerationStage.unprocessed:
     }
@@ -878,7 +883,7 @@ class Param with Annotated implements Element<Param> {
 @JsonSerializable(createToJson: false)
 class Field with ClassMember, Annotated implements Element<Field> {
   Field({
-    this.isExcluded = false,
+    this.isIncluded = true,
     this.annotations,
     this.javadoc,
     this.modifiers = const {},
@@ -888,7 +893,7 @@ class Field with ClassMember, Annotated implements Element<Field> {
   });
 
   @JsonKey(includeFromJson: false)
-  bool isExcluded;
+  bool isIncluded;
 
   @JsonKey(includeFromJson: false)
   String? userDefinedName;