+++ pkgs/jnigen/test/stub_test/generate.dart
@@ -38,18 +38,20 @@ Config getConfig() {
   }
 
   return Config(
-    sourcePath: [Uri.directory(javaPath)],
-    classPath: [Uri.directory(javaPath)],
-    classes: ['com.example.A', 'com.example.C'],
-    outputConfig: OutputConfig(
-      dartConfig: DartCodeOutputConfig(
+    input: Input(
+      sourcePath: [Uri.directory(javaPath)],
+      classPath: [Uri.directory(javaPath)],
+      classes: ['com.example.A', 'com.example.C'],
+    ),
+    output: Output(
+      dart: DartCodeOutput(
         path: Uri.file(dartPath),
         structure: OutputStructure.singleFile,
       ),
+      generateStubs: true,
+      preamble: preamble,
     ),
-    generateStubs: true,
     logLevel: Level.INFO,
-    preamble: preamble,
   );
 }